BE/requirements.txt
numpy==2.3.5
pandas==2.2.3
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.3
elasticsearch==8.12.0
pillow==11.1.0
botocore==1.35.15
ibm-cos-sdk==2.12.0
ibm-watsonx-ai==1.3.28
ibm-cos-sdk==2.12.0
sentence-transformers==3.0.0
google-genai==1.52.0
groq==0.37.1
